refactor: rename new component
